+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Next
+
+-   feature: Allow user to implement custom toggle repository
+-   feat: Drop node 6 support (EOL)
+
 ## 3.2.4
 
 -   feat: add timeout option for http requests (#115)